HumanReadableDurationFormat/src/test/java/TimeFormatterTest.java
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

public class TimeFormatterTest {
    @Test
    public void exampleTests() {
        assertEquals("now", TimeFormatter.formatDuration(0));
        assertEquals("1 second", TimeFormatter.formatDuration(1));
        assertEquals("1 minute and 2 seconds", TimeFormatter.formatDuration(62));
        assertEquals("2 minutes", TimeFormatter.formatDuration(120));
        assertEquals("1 hour", TimeFormatter.formatDuration(3600));
        assertEquals("1 hour, 1 minute and 2 seconds", TimeFormatter.formatDuration(3662));
        assertEquals("1 year, 32 days, 18 hours, 57 minutes and 17 seconds", TimeFormatter.formatDuration(34369037));
        assertEquals("1 year and 1 second", TimeFormatter.formatDuration(31536001));
    }
}
